+++ src/com/rommelrico/view/EmailDetailsLayout.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.web.WebView?>
 
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.rommelrico.EmailDetailsController">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.rommelrico.controller.EmailDetailsController">
    <children>
       <WebView fx:id="webView" layoutX="100.0" layoutY="82.0" prefHeight="200.0" prefWidth="200.0" AnchorPane.bottomAnchor="10.0" AnchorPane.leftAnchor="10.0" AnchorPane.rightAnchor="10.0" AnchorPane.topAnchor="70.0" />
       <Label fx:id="subjectLabel" layoutX="14.0" layoutY="22.0" text="Subject" AnchorPane.leftAnchor="10.0" AnchorPane.topAnchor="10.0" />
